requirements.txt
django==5.0.6
ruff==0.12.0 # dev
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
yfinance==0.2.64
scikit-learn==1.7.0
tensorflow==2.19.0
matplotlib==3.10.3
ipython==9.4.0 # dev
psycopg2-binary==2.9.10
python-telegram-bot==22.1
python-dotenv==1.1.0
celery==5.4.0
redis==5.0.1
whitenoise==6.6.0
django-mssql-backend==2.8.1
mysqlclient==2.2.4 #dev
mssql-django==1.5
gunicorn==23.0.0 # production
stripe==9.12.0
